requirements.txt
Jinja2==2.8
Markdown==2.6.6
MarkupSafe==0.23
Pygments==2.1.3
PyYAML==3.12
